MLOps_Airflow/requirements.txt
scikit-learn
PyYAML
pandas